LPS-66938 Move layout set prototype data handlers to the proper package
