+++ modules/apps/web-experience/layout/layout-set-prototype-impl/src/main/java/com/liferay/layout/set/prototype/internal/exportimport/data/handler/SiteTemplateSettingsPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.set.prototype.lar;
+package com.liferay.layout.set.prototype.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.PortletDataHandler;
 import com.liferay.layout.set.prototype.constants.LayoutSetPrototypePortletKeys;
