--- modules/apps/web-experience/layout/layout-set-prototype-impl/src/main/resources/com/liferay/layout/set/prototype/lar/packageinfo
@@ -1,1 +1,0 @@
-version 0.0.0
